fixed some feedback
