+++ components/templates/admin/news/list.tsx
@@ -9,15 +9,16 @@ import { firebase } from "../../../../firebase";
 import "firebase/firestore";
 import { useRouter } from 'next/router';
 import { DeleteNewsById } from '../../../../firestore/news/deleteNewsById';
+import { useEffect } from 'react';
 
 export default function NewsListTemplate({ newsList }: { newsList: News[] }) {
   const router = useRouter();
 
-  async function fDeleteNewsById(id: string) {
-
-    await DeleteNewsById(id);
+  const handleDelete = (e, id:string) => {
+    e.preventDefault();
+    DeleteNewsById(id);
     alert("Document successfully deleted!"+id);
-    router.reload();
+    router.replace('/admin/news/list');
    
   }
     return (
@@ -67,7 +68,7 @@ export default function NewsListTemplate({ newsList }: { newsList: News[] }) {
               </TD>
               <TD>
               <Link href="#" passHref>
-                <Button onClick={(e) => fDeleteNewsById(`${item.id}`)}>Delete</Button>
+                <Button onClick={(e) => {handleDelete(e, `${item.id}`)}}>Delete</Button>
               </Link>
               </TD>
             </TR>
