+++ firestore/news/deleteNewsById.ts
@@ -3,6 +3,5 @@ import { firebase } from "../../firebase";
 import "firebase/firestore";
 
 export const DeleteNewsById = async(id: string): Promise<void> => {
-    const db = firebase.firestore();
-    await db.collection("news").doc(id).delete();
+    await firebase.firestore().collection("news").doc(id).delete();
 }
