+++ firestore/news/insertNews.ts
@@ -3,9 +3,8 @@ import "firebase/firestore";
 import {News} from "../../models/news";
 
 export const InsertNews = async(news: News): Promise<string> => {  
-    const db = firebase.firestore();
 
-    const doc = await db.collection("news").add(news);
+    const doc = await firebase.firestore().collection("news").add(news);
 
     return doc.id;
