+++ firestore/news/updateNews.ts
@@ -8,11 +8,10 @@ import {News} from "../../models/news";
 // async function UpdateNews(news: News): string {  => LỖI :Type 'string' is not a valid async function return type in ES5/ES3 because it does not refer to a Promise-compatible constructor value.
 export const UpdateNews = async(news: News): Promise<void> => {
    
-    const db = firebase.firestore();
     const dataUpdate:News = {
         title: news.title,
         description: news.description
     };
-    await db.collection("news").doc(news.id).update(dataUpdate);
+    await firebase.firestore().collection("news").doc(news.id).update(dataUpdate);
 
 }
